Refactor: New clean repo created
Created a new clean repository to link in the Batchelors thesis I just handed in. I was quite a newbie when managing the old repo and it's history is flooded with large database files that would be incredibly tedious to remove. This is where continuous development will happen to adapt this project to other things. Even though my Batchelors is nearing completion I still have many things that I would like to explore.